Re-structure Triggers service to match the new standard module architecture of Bulo
