+++ bulo-trigger-scheduler/bulo-trigger-scheduler-app/src/main/java/com/maukaim/bulo/trigger/scheduler/app/TriggerSchedulerMicroServiceApplication.java
@@ -1,0 +1,12 @@
+package com.maukaim.bulo.trigger.scheduler.app;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TriggerSchedulerMicroServiceApplication {
+
+    public static void main(String[] args) throws InterruptedException {
+        SpringApplication.run(TriggerSchedulerMicroServiceApplication.class, args);
+    }
+}
