+++ bulo-trigger-scheduler/bulo-trigger-scheduler-app/src/main/java/com/maukaim/bulo/trigger/scheduler/app/beans/PublisherBeansConfig.java
@@ -1,0 +1,16 @@
+package com.maukaim.bulo.trigger.scheduler.app.beans;
+
+import com.maukaim.bulo.triggers.core.DummyTriggerEventPublisher;
+import com.maukaim.bulo.triggers.core.TriggerEventPublisher;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class PublisherBeansConfig {
+
+    @Bean
+    public TriggerEventPublisher getPublisher(){
+        return new DummyTriggerEventPublisher();
+    }
+
+}
