+++ bulo-trigger-scheduler/bulo-trigger-scheduler-app/src/main/java/com/maukaim/bulo/trigger/scheduler/app/deserializers/DeserializerConfig.java
@@ -1,0 +1,21 @@
+package com.maukaim.bulo.trigger.scheduler.app.deserializers;
+
+import com.maukaim.bulo.commons.serialization.CommonMixinsConfig;
+import com.maukaim.bulo.trigger.scheduler.serialization.SchedulerMixinsConfig;
+import com.maukaim.bulo.triggers.serialization.TriggersMixinsConfig;
+import org.springframework.boot.autoconfigure.jackson.Jackson2ObjectMapperBuilderCustomizer;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class DeserializerConfig {
+
+    @Bean
+    public Jackson2ObjectMapperBuilderCustomizer customizer() {
+        return jacksonObjectMapperBuilder -> {
+            jacksonObjectMapperBuilder.mixIns(CommonMixinsConfig.COMMON_SERIALIZATION_JACKSON_MIXIN);
+            jacksonObjectMapperBuilder.mixIns(TriggersMixinsConfig.TRIGGERS_JACKSON_MIXIN);
+            jacksonObjectMapperBuilder.mixIns(SchedulerMixinsConfig.TRIGGER_SCHEDULER_JACKSON_MIXIN);
+        };
+    }
+}
