+++ bulo-trigger-scheduler/bulo-trigger-scheduler-app/src/main/resources/application.yml
