+++ bulo-trigger-scheduler/bulo-trigger-scheduler-core/src/main/java/com/maukaim/bulo/trigger/scheduler/core/CronExecutorService.java
@@ -1,0 +1,7 @@
+package com.maukaim.bulo.trigger.scheduler.core;
+
+import java.util.concurrent.ScheduledFuture;
+
+public interface CronExecutorService {
+    ScheduledFuture<?> schedule(Runnable task, String cronExpression);
+}
