+++ bulo-trigger-scheduler/bulo-trigger-scheduler-serialization-module/src/main/java/com/maukaim/bulo/trigger/scheduler/serialization/mixins/ScheduleTriggerConfigMixIn.java
@@ -1,0 +1,12 @@
+package com.maukaim.bulo.trigger.scheduler.serialization.mixins;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.maukaim.bulo.triggers.io.TriggerId;
+
+public class ScheduleTriggerConfigMixIn {
+
+    @JsonCreator
+    public ScheduleTriggerConfigMixIn(@JsonProperty("triggerId") TriggerId triggerId,
+                                      @JsonProperty("cronExpression") String cronExpression){}
+}
