--- bulo-trigger-scheduler/src/main/java/com/maukaim/bulo/trigger/scheduler/TriggerConfig.java
@@ -1,15 +1,0 @@
-package com.maukaim.bulo.trigger.scheduler;
-
-import com.maukaim.bulo.trigger.core.DummyTriggerEventPublisher;
-import com.maukaim.bulo.trigger.core.TriggerEventPublisher;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class TriggerConfig {
-
-    @Bean
-    public TriggerEventPublisher getPublisher(){
-        return new DummyTriggerEventPublisher();
-    }
-}
