--- bulo-trigger-scheduler/src/main/java/com/maukaim/bulo/trigger/scheduler/TriggerSchedulerApplication.java
@@ -1,12 +1,0 @@
-package com.maukaim.bulo.trigger.scheduler;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TriggerSchedulerApplication {
-
-    public static void main(String[] args) throws InterruptedException {
-        SpringApplication.run(TriggerSchedulerApplication.class, args);
-    }
-}
