--- bulo-triggers-core/src/main/java/com/maukaim/bulo/trigger/core/TriggerEventPublisher.java
@@ -1,7 +1,0 @@
-package com.maukaim.bulo.trigger.core;
-
-import com.maukaim.bulo.triggers.api.BasicTriggerEvent;
-
-public interface TriggerEventPublisher{
-    void publish(BasicTriggerEvent event);
-}
