--- bulo-triggers-core/src/main/java/com/maukaim/bulo/trigger/core/TriggerService.java
@@ -1,13 +1,0 @@
-package com.maukaim.bulo.trigger.core;
-
-
-
-import com.maukaim.bulo.commons.models.FlowStageId;
-
-import java.util.Set;
-
-public interface TriggerService<TC> {
-
-    TC setTrigger(TC triggerConfig);
-    boolean removeTrigger(String flowId, Set<FlowStageId> stageIds);
-}
