+++ bulo-triggers-core/src/main/java/com/maukaim/bulo/triggers/core/TriggerEventPublisher.java
@@ -1,0 +1,7 @@
+package com.maukaim.bulo.triggers.core;
+
+import com.maukaim.bulo.triggers.io.out.BasicTriggerEvent;
+
+public interface TriggerEventPublisher{
+    void publish(BasicTriggerEvent event);
+}
