+++ bulo-triggers-serialization-module/pom.xml
@@ -9,12 +9,7 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>bulo-triggers-core</artifactId>
-    <properties>
-        <maven.compiler.source>16</maven.compiler.source>
-        <maven.compiler.target>16</maven.compiler.target>
-        <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-    </properties>
+    <artifactId>bulo-triggers-serialization-module</artifactId>
 
     <dependencies>
         <dependency>
@@ -22,5 +17,13 @@
             <artifactId>bulo-triggers-io-module</artifactId>
             <version>${revision}</version>
         </dependency>
+
+        <dependency>
+            <groupId>com.fasterxml.jackson.core</groupId>
+            <artifactId>jackson-annotations</artifactId>
+            <version>2.11.4</version>
+            <scope>provided</scope>
+        </dependency>
     </dependencies>
+
 </project>
